contando execucoes insetionsort

+++ source/heap_sort.c
@@ -1,12 +1,12 @@
 #include "heap_sort.h"
 
-void swap(int *a, int *b) {
-    int temp = *a;
+void swap(float *a, float *b) {
+    float temp = *a;
     *a = *b;
     *b = temp;
 }
 
-void heapify(int* arr, int n, int i) {
+void heapify(float* arr, int n, int i) {
     int largest = i;
     int left = 2 * i + 1;
     int right = 2 * i + 2;
@@ -23,7 +23,7 @@ void heapify(int* arr, int n, int i) {
     }
 }
 
-int* heap_sort(int* arr, int n) {
+float* heap_sort(float* arr, int n) {
     for (int i = n / 2 - 1; i >= 0; i--) {
         heapify(arr, n, i);
     }
