+++ source/merge_sort.h
@@ -7,8 +7,8 @@
 #ifndef MERGESORT_H
 #define MERGESORT_H
 
-void merge(int* arr, int l, int m, int r);
-int* merge_sort(int* arr, int l, int r);
+void merge(float* arr, int l, int m, int r);
+float* merge_sort(float* arr, int l, int r);
 
 void merge_string(char** words, int l, int m, int r);
 char** merge_sort_string(char** words, int l, int r);
